# docs/specs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "_KbResolveBody": {
        "properties": {
          "url": {
            "default": "",
            "maxLength": 2048,
            "title": "Url",
            "type": "string"
          }
        },
        "title": "_KbResolveBody",
        "type": "object"
      },
      "_KbSourcesBody": {
        "items": {
          "additionalProperties": true,
          "type": "object"
        },
        "maxItems": 500,
        "title": "_KbSourcesBody",
        "type": "array"
      },
...
    }
  },
  "info": {
    "description": "Private local memory — notes, mails, messages, and docs",
    "title": "Estormi MCP",
    "version": "0.0.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/knowledge/resolve": {
      "post": {
        "description": "Inspect a pasted URL and return a pre-filled knowledge-source draft.\n\nDeduces the source type from the URL, fetches a human label from the\nchannel/feed metadata, and guesses a kind. Every field is editable in the\nUI — this endpoint only supplies sensible defaults.",
        "operationId": "resolve_knowledge_source_api_knowledge_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_KbResolveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Knowledge Source"
      }
    },
...
    "/api/knowledge/sources": {
      "get": {
        "operationId": "get_knowledge_sources_api_knowledge_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Knowledge Sources"
      },
      "put": {
        "operationId": "put_knowledge_sources_api_knowledge_sources_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_KbSourcesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Knowledge Sources"
      }
    },
...
    "/api/model/delete": {
      "post": {
        "description": "Delete a downloaded model GGUF to reclaim disk.\n\nA model is binary in the UI — installed or not — so the catalog offers\nDownload when absent and Delete when present. ``tier`` is validated against\nthe catalog (no path traversal: the path is derived from the tier, never\nfrom caller input). If the model being deleted is the one currently\nresident in the sidecar, it is unloaded first so inference isn't left\npointing at an unlinked file.",
        "operationId": "model_delete_api_model_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Model Delete"
      }
    },
...
  }
}